Adding files for handle iframe
