+++ Handle_iframe_in_Selenium/pom.xml
@@ -39,6 +39,25 @@
             <version>4.13.2</version>
             <scope>test</scope>
         </dependency>
+
+    <dependency>
+        <groupId>net.bytebuddy</groupId>
+        <artifactId>byte-buddy</artifactId>
+        <version>LATEST</version>
+    </dependency>
+
+        <dependency>
+            <groupId>com.google.guava</groupId>
+            <artifactId>guava</artifactId>
+            <version>23.0</version>
+        </dependency>
+        <dependency>
+            <groupId>com.google.code.gson</groupId>
+            <artifactId>gson</artifactId>
+            <version>2.8.2</version>
+        </dependency>
+
     </dependencies>
 
 </project>
+
