+++ Handle_iframe_in_Selenium/src/main/resources/data.properties
@@ -1,0 +1,3 @@
+browser=chrome
+url =https://developer.mozilla.org/en-US/docs/Learn/HTML/Multimedia_and_embedding/Other_embedding_technologies
+
