--- src/main/resources/data.properties
@@ -1,2 +1,0 @@
-browser=chrome
-Url1=https://www.youtube.com
